move stale job purging to tasks/general.py
