+++ redash/tasks/__init__.py
@@ -1,4 +1,4 @@
-from .general import record_event, version_check, send_mail, sync_user_details
+from .general import record_event, version_check, send_mail, sync_user_details, purge_failed_jobs
 from .queries import (QueryTask, enqueue_query, execute_query, refresh_queries,
                       refresh_schemas, cleanup_query_results, empty_schedules)
 from .alerts import check_alerts_for_query
